requirements.txt
Package         Version
--------------- -------
asgiref         3.5.2
Django          4.1.2
django-ckeditor 6.5.1
django-js-asset 2.0.0
django-resized  1.0.2
Pillow          9.2.0
pip             22.2.2
setuptools      63.2.0
sqlparse        0.4.3
whitenoise      6.2.0